refactor: remove long
long is 32-bits even on 64-bit windows which makes the type suboptimal
for a codebase meant to be cross-platform.

+++ src/nvim/getchar.c
@@ -2695,9 +2695,9 @@ static int vgetorpeek(bool advance)
             // blocking wait
             wait_time = -1L;
           } else if (keylen == KEYLEN_PART_KEY && p_ttm >= 0) {
-            wait_time = p_ttm;
+            wait_time = (long)p_ttm;
           } else {
-            wait_time = p_tm;
+            wait_time = (long)p_tm;
           }
         }
